+++ lib/events/system_message.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Events
+  class SystemMessage < Base
+    TYPE = "system_message"
+
+    def type
+      TYPE
+    end
+  end
+end